requirements.txt
asgiref==3.3.1
boto3==1.16.51
botocore==1.19.51
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
jmespath==0.10.0
Pillow==8.1.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
